requirements.txt
feedparser
youtube-transcript-api
openai
python-dotenv
PyYAML
jinja2
